disabled chunks + dynamically load lottie animations
