+++ theme/animations/AddingStopLossAnimation.tsx
@@ -1,0 +1,20 @@
+import Lottie from 'react-lottie'
+import { Box } from 'theme-ui'
+import addingStopLossAnimation from 'theme/lottie/addingStopLossAnimation.json'
+
+const addingStopLossAnimationOptions = {
+  loop: true,
+  autoplay: true,
+  animationData: addingStopLossAnimation,
+  rendererSettings: {
+    preserveAspectRatio: 'xMaxYMid slice',
+  },
+}
+
+export default function AddingStopLossAnimationStatic() {
+  return (
+    <Box mb={2}>
+      <Lottie options={addingStopLossAnimationOptions} height={180} width={280} />
+    </Box>
+  )
+}
